+++ packages/core/src/utils/webauthn-shared.ts
